aggiornamento classe DataSet, WeatherUtils, e correzione sptingbootapplication

+++ EsameProgrammazione/src/main/java/it/univpm/EsameProgrammazione/Utilities/DataSet.java
@@ -1,14 +1,11 @@
 package it.univpm.EsameProgrammazione.Utilities;
-import it.univpm.EsameProgrammazione.Model.*;
 
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.util.Vector;
-
+import it.univpm.EsameProgrammazione.Model.Weather;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+
+import java.io.*;
+import java.util.Vector;
 
 /*
  * Classe che si occupa della gestione del DataSet
@@ -25,17 +22,23 @@ public class DataSet{
 	
 	public void addWeather(Weather objectWeather) {
 		weathers.add(objectWeather);
+
+
 		weathers.toString();
 	}
-	
+
+	File name;
 	@Scheduled(fixedRate = 30000)
 	public void saveInDataSet() throws IOException
 	{
 		try {
-			FileOutputStream fos = new FileOutputStream("dataSet.txt");
-		    ObjectOutputStream oos = new ObjectOutputStream(fos);
-		    oos.writeObject(this.weathers);
-		    oos.close();
+			name = new File("dataSet.txt");
+			if (name.exists()) {
+				FileOutputStream fos = new FileOutputStream("dataSet.txt");
+				ObjectOutputStream oos = new ObjectOutputStream(fos);
+				oos.writeObject(this.weathers);
+				oos.close();
+			}
 		}
 		catch(FileNotFoundException e) {
 			e.printStackTrace();
@@ -44,5 +47,6 @@ public class DataSet{
 			e.printStackTrace();
 		}
 	}
-	
+
+
 }
